85: run on .NetCore
the netcore version works on Windows 64 only,
Linux and MacOS coming soon
