+++ .vs/restore.dg
@@ -1,0 +1,1 @@
+#:D:\projects\Espresso\EspressoCore\EspressoCore.xproj
