--- EspressoCore/EspressoCore.csproj
@@ -1,83 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
-  <PropertyGroup>
-    <MinimumVisualStudioVersion>10.0</MinimumVisualStudioVersion>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{90183F26-15B6-45E6-A5F0-6E287EEE406A}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>EspressoCore</RootNamespace>
-    <AssemblyName>EspressoCore</AssemblyName>
-    <DefaultLanguage>en-US</DefaultLanguage>
-    <FileAlignment>512</FileAlignment>
-    <ProjectTypeGuids>{786C830F-07A1-408B-BD7F-6EE04809D6DB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
-    <TargetFrameworkProfile>Profile111</TargetFrameworkProfile>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <ItemGroup>
-    <!-- A reference to the entire .NET Framework is automatically included -->
-    <ProjectReference Include="..\Espresso.Interfaces\Espresso.Interfaces.csproj">
-      <Project>{170910c7-e88f-47d8-89aa-fab3b0e0ff9c}</Project>
-      <Name>Espresso.Interfaces</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="0_Start\JsBridge.cs" />
-    <Compile Include="0_Start\ScriptInterfaces.cs" />
-    <Compile Include="1_Core\BoundWeakDelegate.cs" />
-    <Compile Include="1_Core\IKeepAliveStore.cs" />
-    <Compile Include="1_Core\JsContext.cs" />
-    <Compile Include="1_Core\JsContext.Dynamic.cs" />
-    <Compile Include="1_Core\JsContext.NativeMethods.cs" />
-    <Compile Include="1_Core\JsConvert.cs" />
-    <Compile Include="1_Core\JsEngine.cs" />
-    <Compile Include="1_Core\JsEngine.NativeMethods.cs" />
-    <Compile Include="1_Core\JsEngineStats.cs" />
-    <Compile Include="1_Core\JsError.cs" />
-    <Compile Include="1_Core\JsException.cs" />
-    <Compile Include="1_Core\JsExecutionTimedOutException.cs" />
-    <Compile Include="1_Core\JsFunction.cs" />
-    <Compile Include="1_Core\JsInteropException.cs" />
-    <Compile Include="1_Core\JsObject.Dynamic.cs" />
-    <Compile Include="1_Core\JsObjectMarshalType.cs" />
-    <Compile Include="1_Core\JsScript.cs" />
-    <Compile Include="1_Core\JsValue.cs" />
-    <Compile Include="1_Core\JsValueType.cs" />
-    <Compile Include="1_Core\KeepAliveDictionaryStore.cs" />
-    <Compile Include="1_Core\WeakDelegate.cs" />
-    <Compile Include="3_MiniBridge\DelegateHolders.cs" />
-    <Compile Include="3_MiniBridge\JsTypeDefinitionBuilder.cs" />
-    <Compile Include="3_MiniBridge\NativeV8JsInterOp.cs" />
-    <Compile Include="Extension.cs" />
-    <Compile Include="HandleRef.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
-  </ItemGroup>
-  <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
-  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
-       Other similar extension points exist, see Microsoft.Common.targets.
-  <Target Name="BeforeBuild">
-  </Target>
-  <Target Name="AfterBuild">
-  </Target>
-  -->
-</Project>
