+++ Net20/Espresso.Interfaces/Espresso.Interfaces.csproj
@@ -35,12 +35,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\..\Espresso.Interfaces\IScriptable.cs">
-      <Link>IScriptable.cs</Link>
-    </Compile>
+    <Folder Include="Properties\" />
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="Properties\" />
+    <Compile Include="..\..\EspressoCore\IScriptable.cs">
+      <Link>IScriptable.cs</Link>
+    </Compile>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
