+++ Portable/EspressoCore/EspressoCore.csproj
@@ -1,0 +1,140 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
+  <PropertyGroup>
+    <MinimumVisualStudioVersion>10.0</MinimumVisualStudioVersion>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProjectGuid>{90183F26-15B6-45E6-A5F0-6E287EEE406A}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <AppDesignerFolder>Properties</AppDesignerFolder>
+    <RootNamespace>EspressoCore</RootNamespace>
+    <AssemblyName>EspressoCore</AssemblyName>
+    <DefaultLanguage>en-US</DefaultLanguage>
+    <FileAlignment>512</FileAlignment>
+    <ProjectTypeGuids>{786C830F-07A1-408B-BD7F-6EE04809D6DB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
+    <TargetFrameworkProfile>Profile111</TargetFrameworkProfile>
+    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+    <DebugSymbols>true</DebugSymbols>
+    <DebugType>full</DebugType>
+    <Optimize>false</Optimize>
+    <OutputPath>bin\Debug\</OutputPath>
+    <DefineConstants>DEBUG;TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <DebugType>pdbonly</DebugType>
+    <Optimize>true</Optimize>
+    <OutputPath>bin\Release\</OutputPath>
+    <DefineConstants>TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <ItemGroup>
+    <!-- A reference to the entire .NET Framework is automatically included -->
+    <ProjectReference Include="..\Espresso.Interfaces\Espresso.Interfaces.csproj">
+      <Project>{170910c7-e88f-47d8-89aa-fab3b0e0ff9c}</Project>
+      <Name>Espresso.Interfaces</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="..\..\EspressoCore\0_Start\JsBridge.cs">
+      <Link>0_Start\JsBridge.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\0_Start\ScriptInterfaces.cs">
+      <Link>0_Start\ScriptInterfaces.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\BoundWeakDelegate.cs">
+      <Link>1_Core\BoundWeakDelegate.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\IKeepAliveStore.cs">
+      <Link>1_Core\IKeepAliveStore.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsContext.cs">
+      <Link>1_Core\JsContext.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsContext.Dynamic.cs">
+      <Link>1_Core\JsContext.Dynamic.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsContext.NativeMethods.cs">
+      <Link>1_Core\JsContext.NativeMethods.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsConvert.cs">
+      <Link>1_Core\JsConvert.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsEngine.cs">
+      <Link>1_Core\JsEngine.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsEngine.NativeMethods.cs">
+      <Link>1_Core\JsEngine.NativeMethods.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsEngineStats.cs">
+      <Link>1_Core\JsEngineStats.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsError.cs">
+      <Link>1_Core\JsError.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsException.cs">
+      <Link>1_Core\JsException.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsExecutionTimedOutException.cs">
+      <Link>1_Core\JsExecutionTimedOutException.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsFunction.cs">
+      <Link>1_Core\JsFunction.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsInteropException.cs">
+      <Link>1_Core\JsInteropException.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsObject.Dynamic.cs">
+      <Link>1_Core\JsObject.Dynamic.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsObjectMarshalType.cs">
+      <Link>1_Core\JsObjectMarshalType.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsScript.cs">
+      <Link>1_Core\JsScript.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsValue.cs">
+      <Link>1_Core\JsValue.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\JsValueType.cs">
+      <Link>1_Core\JsValueType.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\KeepAliveDictionaryStore.cs">
+      <Link>1_Core\KeepAliveDictionaryStore.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\1_Core\WeakDelegate.cs">
+      <Link>1_Core\WeakDelegate.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\3_MiniBridge\DelegateHolders.cs">
+      <Link>3_MiniBridge\DelegateHolders.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\3_MiniBridge\JsTypeDefinitionBuilder.cs">
+      <Link>3_MiniBridge\JsTypeDefinitionBuilder.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\3_MiniBridge\NativeV8JsInterOp.cs">
+      <Link>3_MiniBridge\NativeV8JsInterOp.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\Extension.cs">
+      <Link>Extension.cs</Link>
+    </Compile>
+    <Compile Include="..\..\EspressoCore\HandleRef.cs">
+      <Link>HandleRef.cs</Link>
+    </Compile>
+    <Compile Include="Properties\AssemblyInfo.cs" />
+  </ItemGroup>
+  <ItemGroup />
+  <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
+  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
+       Other similar extension points exist, see Microsoft.Common.targets.
+  <Target Name="BeforeBuild">
+  </Target>
+  <Target Name="AfterBuild">
+  </Target>
+  -->
+</Project>
