+++ TestEspressoCore/Program.cs
@@ -1,0 +1,90 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Reflection;
+namespace TestEspressoCore
+{
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+            //VroomJs.JsBridge.LoadV8(@"D:\projects\Espresso\build\Debug\libespr.dll");
+            //D:\projects\Espresso\TestEspressoCore\bin\Debug\netcoreapp1.0
+            VroomJs.JsBridge.LoadV8(@"D:\projects\Espresso\build\Debug\netcoreapp1.0\xlibespr.dll");
+            //if (System.IO.File.Exists(@"libespr.dll"))
+            //{
+            //    VroomJs.JsBridge.LoadV8(@"D:\projects\Espresso\Release\libespr.dll");
+            //}
+            //----------------------------
+            //prepare test cases 
+            var testList = GetTestInfoList();
+            //----------------------------
+            testDictionary = new Dictionary<string, TestInfo>();
+            foreach (TestInfo t in testList)
+            {
+                testDictionary.Add(t.Choice, t);
+            }
+
+            AGAIN:
+            //----------------------------------------------
+            Menu(testList);
+            Console.WriteLine("");
+            Console.WriteLine("type 0 to exit");
+            //----------------------------------------------
+            string num = Console.ReadLine();
+            Console.WriteLine("running ...");
+            Console.WriteLine("");
+            if (num == "0")
+            {
+                return;
+            }
+            TestInfo selectedTest;
+            if (testDictionary.TryGetValue(num, out selectedTest))
+            {
+                //found test
+                selectedTest.TestMethod.Invoke(num, new object[0]);
+            }
+            else
+            {
+                Console.WriteLine("---[not found, Please press choice (1,2 etc)]---");
+            }
+            Console.WriteLine("------");
+            Console.WriteLine("");
+            goto AGAIN;
+        }
+
+
+        static List<TestInfo> GetTestInfoList()
+        {
+            List<TestInfo> testList = new List<TestInfo>();
+            Type t = typeof(Test1);
+            var mets = t.GetMethods(System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.NonPublic);
+            foreach (var met in mets)
+            {
+                var found = met.GetCustomAttributes(typeof(TestAttribute), false) as TestAttribute[];
+                if (found != null && found.Length > 0)
+                {
+                    TestInfo testInfo = new TestInfo()
+                    {
+                        Choice = found[0].Choice,
+                        Name = found[0].Name,
+                        TestMethod = met
+                    };
+                    testList.Add(testInfo);
+                }
+            }
+            return testList;
+        }
+
+        static Dictionary<string, TestInfo> testDictionary;
+        static void Menu(List<TestInfo> testList)
+        {
+            Console.WriteLine("Espresso");
+            Console.WriteLine("Select test case, and press Enter");
+            Console.WriteLine("[0] Exit");
+            foreach (TestInfo t in testList)
+            {
+                Console.WriteLine("[" + t.Choice + "] " + t.Name);
+            }
+        }
+    }
+}
