+++ TestEspressoCore/Tests/TestInfo.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.IO;
+using VroomJs;
+
+namespace TestEspressoCore
+{
+    class TestInfo
+    {
+
+        public TestInfo()
+        {
+
+        }
+        public string Name { get; set; }
+        public string Choice { get; set; }
+        public System.Reflection.MethodInfo TestMethod { get; set; }
+    }
+}
