+++ TestEspressoCore/project.json
@@ -1,0 +1,21 @@
+{
+  "buildOptions": {
+    "emitEntryPoint": true,
+    "allowUnsafe": true,
+    "platform": "anycpu32bitpreferred",
+    "debugType": "full"
+  },
+  "dependencies": {  
+    "EspressoCore": "1.0.0-*",
+    "Microsoft.NETCore.App": {
+      "type": "platform",
+      "version": "1.0.0"
+    }
+  },
+  "frameworks": {
+    "netcoreapp1.0": {
+      "imports": "dnxcore50"
+    }
+  },
+  "version": "1.0.0-*"
+}
