+++ VroomJsConsole/VroomJsConsole.csproj
@@ -36,17 +36,24 @@
     <Reference Include="System" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="..\TestEspressoCore\Tests\Test1.cs">
+      <Link>Tests\Test1.cs</Link>
+    </Compile>
+    <Compile Include="..\TestEspressoCore\Tests\TestAttribute.cs">
+      <Link>Tests\TestAttribute.cs</Link>
+    </Compile>
+    <Compile Include="..\TestEspressoCore\Tests\TestInfo.cs">
+      <Link>Tests\TestInfo.cs</Link>
+    </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestAttribute.cs" />
-    <Compile Include="TestInfo.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Espresso.Interfaces\Espresso.Interfaces.csproj">
+    <ProjectReference Include="..\Portable\Espresso.Interfaces\Espresso.Interfaces.csproj">
       <Project>{170910c7-e88f-47d8-89aa-fab3b0e0ff9c}</Project>
       <Name>Espresso.Interfaces</Name>
     </ProjectReference>
-    <ProjectReference Include="..\EspressoCore\EspressoCore.csproj">
+    <ProjectReference Include="..\Portable\EspressoCore\EspressoCore.csproj">
       <Project>{90183f26-15b6-45e6-a5f0-6e287eee406a}</Project>
       <Name>EspressoCore</Name>
     </ProjectReference>
