+++ libvroomjs/libvroomjs.sln
@@ -9,12 +9,18 @@ Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x86 = Debug|x86
 		Release|x86 = Release|x86
+		Debug|x64 = Debug|x64
+		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{880EF82B-4FAD-42B1-98A6-09B5835F5FE6}.Debug|x86.ActiveCfg = Debug|Win32
 		{880EF82B-4FAD-42B1-98A6-09B5835F5FE6}.Debug|x86.Build.0 = Debug|Win32
 		{880EF82B-4FAD-42B1-98A6-09B5835F5FE6}.Release|x86.ActiveCfg = Release|Win32
 		{880EF82B-4FAD-42B1-98A6-09B5835F5FE6}.Release|x86.Build.0 = Release|Win32
+		{880EF82B-4FAD-42B1-98A6-09B5835F5FE6}.Debug|x64.ActiveCfg = Release|x64
+		{880EF82B-4FAD-42B1-98A6-09B5835F5FE6}.Debug|x64.Build.0 = Release|x64
+		{880EF82B-4FAD-42B1-98A6-09B5835F5FE6}.Release|x64.ActiveCfg = Release|x64
+		{880EF82B-4FAD-42B1-98A6-09B5835F5FE6}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
